chore: replace ui folder with interchain-ui folder
