--- interchain-ui/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/agoric.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>dApp Agoric Basics</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
