--- interchain-ui/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "interchain-ui",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@agoric/react-components": "^0.1.1-dev-8fc28e8.0",
-    "@agoric/ui-components": "^0.3.9-u13.0",
-    "@agoric/web-components": "^0.15.1-dev-09ca26c.0",
-    "buffer": "^6.0.3",
-    "chain-registry": "^1.28.1",
-    "cosmos-kit": "^2.9.0",
-    "daisyui": "^4.7.2",
-    "react": "^18.2.0",
-    "react-daisyui": "^5.0.0",
-    "react-dom": "^18.2.0",
-    "ses": "^1.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.56",
-    "@types/react-dom": "^18.2.19",
-    "@typescript-eslint/eslint-plugin": "^7.0.2",
-    "@typescript-eslint/parser": "^7.0.2",
-    "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.17",
-    "eslint": "^8.56.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
-    "postcss": "^8.4.35",
-    "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.11",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5.2.2",
-    "vite": "^5.1.4"
-  },
-  "prettier": {
-    "trailingComma": "all",
-    "arrowParens": "avoid",
-    "singleQuote": true,
-    "tabWidth": 2,
-    "plugins": [
-      "prettier-plugin-tailwindcss"
-    ]
-  }
-}
