+++ ui/package.json
@@ -1,55 +1,51 @@
 {
-  "name": "agoric-basics-ui",
+  "name": "interchain-ui",
   "private": true,
   "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "test": "vitest spec",
-    "test:e2e": "vitest e2e",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "lint:fix": "yarn lint --fix",
     "preview": "vite preview"
   },
   "dependencies": {
+    "@agoric/react-components": "^0.1.1-dev-8fc28e8.0",
+    "@agoric/ui-components": "^0.3.9-u13.0",
+    "@agoric/web-components": "^0.15.1-dev-09ca26c.0",
+    "buffer": "^6.0.3",
+    "chain-registry": "^1.28.1",
+    "cosmos-kit": "^2.9.0",
+    "daisyui": "^4.7.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-daisyui": "^5.0.0",
+    "react-dom": "^18.2.0",
+    "ses": "^1.2.0"
   },
   "devDependencies": {
-    "@agoric/eventual-send": "^0.14.1",
-    "@agoric/notifier": "^0.6.2",
-    "@agoric/rpc": "0.9.1-dev-f471a83.0",
-    "@agoric/store": "^0.9.2",
-    "@agoric/ui-components": "^0.9.0",
-    "@agoric/web-components": "^0.15.0",
-    "@testing-library/react": "^14.1.2",
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "@vitejs/plugin-react": "^4.0.3",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
+    "@vitejs/plugin-react": "^4.2.1",
+    "autoprefixer": "^10.4.17",
+    "eslint": "^8.56.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.3",
-    "happy-dom": "^13.3.1",
-    "prettier": "^3.2.4",
-    "puppeteer": "^21.9.0",
-    "ses": "^1.3.0",
-    "typescript": "^5.0.2",
-    "vite": "^4.4.5",
-    "vitest": "^1.2.1",
-    "zustand": "^4.4.1"
-  },
-  "resolutions": {
-    "**/ses": "^1.3.0",
-    "**/@agoric/xsnap": "0.14.3-dev-9f085d3.0"
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.35",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.1.4"
   },
   "prettier": {
     "trailingComma": "all",
     "arrowParens": "avoid",
-    "singleQuote": true
+    "singleQuote": true,
+    "tabWidth": 2,
+    "plugins": [
+      "prettier-plugin-tailwindcss"
+    ]
   }
 }
