--- ui/src/App.spec.tsx
@@ -1,25 +1,0 @@
-import './installSesLockdown';
-import { render, screen } from '@testing-library/react';
-import App from './App';
-
-describe('App.tsx', () => {
-  it('renders app title', async () => {
-    render(<App />);
-
-    const titleElement = await screen.findByText(
-      'Items Listed on Agoric Basics',
-      {
-        selector: 'h1',
-      },
-    );
-    expect(titleElement).toBeTruthy();
-  });
-
-  it('renders the wallet connection button', async () => {
-    render(<App />);
-    const buttonEl = await screen.findByRole('button', {
-      name: 'Connect Wallet',
-    });
-    expect(buttonEl).toBeTruthy();
-  });
-});
