+++ ui/src/components/Inventory.tsx
@@ -1,49 +1,61 @@
-import { stringifyAmountValue } from '@agoric/ui-components';
+import { AmountMath } from '@agoric/ertp';
+import { ConnectWalletButton, useAgoric } from '@agoric/react-components';
+import { stringifyAmountValue } from '@agoric/web-components';
+import { usePurse } from '../hooks/usePurse';
+import type { CopyBag } from '../types';
 
-type InventoryProps = {
-  address: string;
-  istPurse: Purse;
-  itemsPurse: Purse;
-};
+const Inventory = () => {
+  const istPurse = usePurse('IST');
+  const ticketsPurse = usePurse('Ticket');
+  const { walletConnection } = useAgoric();
 
-const Inventory = ({ address, istPurse, itemsPurse }: InventoryProps) => (
-  <div className="card">
-    <h3>My Wallet</h3>
-    <div>
-      <div>
-        <small>
-          <code>{address}</code>
-        </small>
-      </div>
-
-      <div style={{ textAlign: 'left' }}>
-        <div>
-          <b>IST: </b>
-          {stringifyAmountValue(
-            istPurse.currentAmount,
-            istPurse.displayInfo.assetKind,
-            istPurse.displayInfo.decimalPlaces,
-          )}
-        </div>
-        <div>
-          <b>Items:</b>
-          {itemsPurse ? (
-            <ul style={{ marginTop: 0, textAlign: 'left' }}>
-              {(itemsPurse.currentAmount.value as CopyBag).payload.map(
-                ([name, number]) => (
-                  <li key={name}>
-                    {String(number)} {name}
-                  </li>
-                ),
-              )}
-            </ul>
-          ) : (
-            'None'
-          )}
+  return (
+    <div className="daisyui-card w-96 bg-base-100 shadow-xl">
+      <div className="daisyui-card-body items-center text-center">
+        <h2 className="daisyui-card-title">My Wallet</h2>
+        <div className="daisyui-card-actions">
+          <div>
+            <ConnectWalletButton className="daisyui-btn daisyui-btn-outline daisyui-btn-primary" />
+            {walletConnection && (
+              <div>
+                <div>
+                  <b>IST: </b>
+                  {istPurse ? (
+                    stringifyAmountValue(
+                      AmountMath.make(
+                        istPurse?.currentAmount.brand,
+                        istPurse?.currentAmount.value,
+                      ),
+                      istPurse.displayInfo.assetKind,
+                      istPurse.displayInfo.decimalPlaces,
+                    )
+                  ) : (
+                    <i>Fetching balance...</i>
+                  )}
+                </div>
+                <div>
+                  <b>Tickets: </b>
+                  {ticketsPurse ? (
+                    <ul>
+                      {(
+                        ticketsPurse.currentAmount.value as CopyBag
+                      ).payload.map(([name, number]) => (
+                        <li key={name}>
+                          {String(number)} {name}
+                        </li>
+                      ))}
+                    </ul>
+                  ) : (
+                    'None'
+                  )}
+                </div>
+              </div>
+            )}
+          </div>
         </div>
       </div>
     </div>
-  </div>
-);
+  );
+};
 
 export { Inventory };
