+++ ui/src/components/mint/MintTickets.tsx
