--- ui/src/index.d.ts
@@ -1,16 +1,0 @@
-interface CopyBag<T = string> {
-  payload: Array<[T, bigint]>;
-}
-
-interface Purse {
-  brand: unknown;
-  brandPetname: string;
-  currentAmount: {
-    brand: unknown;
-    value: bigint | CopyBag;
-  };
-  displayInfo: {
-    decimalPlaces: number;
-    assetKind: unknown;
-  };
-}
