+++ ui/src/vite-env.d.ts
@@ -1,9 +1,1 @@
 /// <reference types="vite/client" />
-
-declare module '@agoric/ui-components' {
-  export const stringifyAmountValue;
-}
-
-declare module '@agoric/store' {
-  export const makeCopyBag;
-}
